+++ recipes/default.rb
@@ -1,7 +1,11 @@
-include_recipe 'apt' if node['platform'] == 'ubuntu'
-
-package 'apt-transport-https'
-package 'bsdtar'
+case node['platform']
+when 'ubuntu'
+  include_recipe 'apt'
+  package 'apt-transport-https'
+  package 'bsdtar'
+  include_recipe 'lxc'
+  include_recipe 'docker::aufs'
+end
 
 if node['docker']['install_type'] == 'source'
   node.set['go']['version'] = '1.1'
@@ -9,7 +13,5 @@ if node['docker']['install_type'] == 'source'
   include_recipe 'git'
 end
 
-include_recipe 'lxc'
-include_recipe 'docker::aufs'
 include_recipe "docker::#{node['docker']['install_type']}"
-include_recipe 'docker::upstart'
+include_recipe "docker::#{node['docker']['init_type']}"