+++ recipes/systemd.rb
@@ -1,0 +1,20 @@
+template '/usr/lib/systemd/system/docker.socket' do
+  source 'docker.socket.erb'
+  mode '0644'
+  owner 'root'
+  group 'root'
+end
+
+template '/usr/lib/systemd/system/docker.service' do
+  source 'docker.service.erb'
+  mode '0644'
+  owner 'root'
+  group 'root'
+  notifies :restart, 'service[docker]', :immediately
+end
+
+service 'docker' do
+  provider Chef::Provider::Service::Systemd
+  supports :status => true, :restart => true, :reload => true
+  action [:start, :enable]
+end